requirements.txt
# Core bot dependencies
python-telegram-bot==20.8
openai==1.12.0

# Database - Unified PostgreSQL for both local and Supabase
asyncpg==0.29.0

# Environment and configuration
python-dotenv==1.0.1

# HTTP requests
httpx==0.26.0
requests==2.31.0

# Optional: Web framework for health endpoints
fastapi==0.109.2
uvicorn==0.27.1

# Data processing
numpy==1.26.4
pandas==2.2.0

# Text processing and NLP utilities
nltk==3.8.1
scikit-learn==1.4.0

# Development dependencies (optional, remove for production)
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.2.0
flake8==7.0.0
mypy==1.8.0

# Logging (enhanced)
structlog==24.1.0
colorama==0.4.6

# Performance monitoring (optional)
psutil==5.9.8

# Security
cryptography==42.0.2

# Additional dependencies that may be needed
typing-extensions==4.9.0
pydantic==2.6.1
anyio==4.2.0
sniffio==1.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.0